web/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.12
itsdangerous==2.2.0
starlette==0.41.3
aiosqlite==0.20.0
Pillow==11.0.0
